# docs/openapi.yml
openapi: 3.0.1
info:
  title: sample-bookstore-restful-api
  description: Sample book store REST API service using Spring Boot.
  version: 23.0.0-SNAPSHOT
servers:
  - url: ''
tags:
  - name: Product category
    description: API for product categories management.
  - name: Geo Country
    description: API for geo countries management.
  - name: Checkout
    description: API for checkout actions.
  - name: Product
    description: API for products management.
  - name: Geo State
    description: API for geo states management.
paths:
  /api/v1/checkout/purchase:
    post:
      tags:
        - Checkout
      operationId: placeOrder
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Purchase'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PurchaseResponse'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ErrorResponseInvalidRequestBodyMeta'
                  - $ref: >-
                      #/components/schemas/ErrorResponseInvalidRequestParameterMeta
        '404':
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '405':
          description: Method Not Allowed
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponseInvalidHttpMethodMeta'
        '415':
          description: Unsupported Media Type
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponseInvalidContentTypeMeta'
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponseUnknownErrorMeta'
  /api/v1/states:
    get:
      tags:
        - Geo State
      description: Retrieve states list.
      operationId: findAll
      parameters:
        - name: page
          in: query
          description: Page number
          required: false
          schema:
            minimum: 0
            type: integer
            format: int32
        - name: size
          in: query
          description: Page size
          required: false
          schema:
            maximum: 200
            minimum: 1
            type: integer
            format: int32
      responses:
        '200':
          description: Success
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageGeoState'
        '400':
          description: Bad request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '405':
          description: Method Not Allowed
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponseInvalidHttpMethodMeta'
        '415':
          description: Unsupported Media Type
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponseInvalidContentTypeMeta'
        '500':
          description: Internal error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/states/{id}:
    get:
      tags:
        - Geo State
      description: Retrieve state by id.
      operationId: findById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/GeoState'
        '400':
          description: Bad request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Entity not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '405':
          description: Method Not Allowed
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponseInvalidHttpMethodMeta'
        '415':
          description: Unsupported Media Type
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponseInvalidContentTypeMeta'
        '500':
          description: Internal error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/products:
    get:
      tags:
        - Product
      description: Retrieve products list.
      operationId: findAll_1
      parameters:
        - name: name
          in: query
          required: false
          schema:
            type: string
        - name: page
          in: query
          description: Page number
          required: false
          schema:
            minimum: 0
            type: integer
            format: int32
        - name: size
          in: query
          description: Page size
          required: false
          schema:
            maximum: 200
            minimum: 1
            type: integer
            format: int32
      responses:
        '200':
          description: Success
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageProduct'
        '400':
          description: Bad request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '405':
          description: Method Not Allowed
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponseInvalidHttpMethodMeta'
        '415':
          description: Unsupported Media Type
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponseInvalidContentTypeMeta'
        '500':
          description: Internal error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/products/{id}:
    get:
      tags:
        - Product
      description: Retrieve product by id.
      operationId: findById_1
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Bad request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Entity not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '405':
          description: Method Not Allowed
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponseInvalidHttpMethodMeta'
        '415':
          description: Unsupported Media Type
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponseInvalidContentTypeMeta'
        '500':
          description: Internal error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/product-categories:
    get:
      tags:
        - Product category
      description: Retrieve categories list.
      operationId: findAll_2
      responses:
        '200':
          description: Success
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductCategory'
        '400':
          description: Bad request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '405':
          description: Method Not Allowed
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponseInvalidHttpMethodMeta'
        '415':
          description: Unsupported Media Type
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponseInvalidContentTypeMeta'
        '500':
          description: Internal error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/product-categories/{id}:
    get:
      tags:
        - Product category
      description: Retrieve product category by id.
      operationId: findById_2
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProductCategory'
        '400':
          description: Bad request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Entity not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '405':
          description: Method Not Allowed
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponseInvalidHttpMethodMeta'
        '415':
          description: Unsupported Media Type
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponseInvalidContentTypeMeta'
        '500':
          description: Internal error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/product-categories/{id}/products:
    get:
      tags:
        - Product category
      description: Retrieve products in specified category.
      operationId: findProductsByCategoryId
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: page
          in: query
          description: Page number
          required: false
          schema:
            minimum: 0
            type: integer
            format: int32
        - name: size
          in: query
          description: Page size
          required: false
          schema:
            maximum: 200
            minimum: 1
            type: integer
            format: int32
      responses:
        '200':
          description: Success
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageProduct'
        '400':
          description: Bad request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '405':
          description: Method Not Allowed
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponseInvalidHttpMethodMeta'
        '415':
          description: Unsupported Media Type
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponseInvalidContentTypeMeta'
        '500':
          description: Internal error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/countries:
    get:
      tags:
        - Geo Country
      description: Retrieve countries list.
      operationId: findAll_3
      responses:
        '200':
          description: Success
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GeoCountry'
        '400':
          description: Bad request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '405':
          description: Method Not Allowed
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponseInvalidHttpMethodMeta'
        '415':
          description: Unsupported Media Type
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponseInvalidContentTypeMeta'
        '500':
          description: Internal error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/countries/{id}:
    get:
      tags:
        - Geo Country
      description: Retrieve geo country by id.
      operationId: findById_3
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/GeoCountry'
        '400':
          description: Bad request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Entity not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '405':
          description: Method Not Allowed
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponseInvalidHttpMethodMeta'
        '415':
          description: Unsupported Media Type
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponseInvalidContentTypeMeta'
        '500':
          description: Internal error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/countries/{id}/states:
    get:
      tags:
        - Geo Country
      description: Retrieve states in specified country.
      operationId: findStatesByCountryId
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GeoState'
        '400':
          description: Bad request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '405':
          description: Method Not Allowed
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponseInvalidHttpMethodMeta'
        '415':
          description: Unsupported Media Type
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponseInvalidContentTypeMeta'
        '500':
          description: Internal error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
    ErrorResponse:
      type: object
      properties:
        errors:
          type: array
          items:
            $ref: '#/components/schemas/Error'
    ErrorInvalidRequestParameterMeta:
      type: object
      properties:
        code:
          type: string
        detail:
          type: string
        meta:
          $ref: '#/components/schemas/InvalidRequestParameterMeta'
    ErrorResponseInvalidRequestParameterMeta:
      type: object
      properties:
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ErrorInvalidRequestParameterMeta'
    InvalidRequestParameterMeta:
      type: object
      properties:
        parameter:
          type: string
        invalidValue:
          type: string
    ErrorResponseUnknownErrorMeta:
      type: object
      properties:
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ErrorUnknownErrorMeta'
    ErrorUnknownErrorMeta:
      type: object
      properties:
        code:
          type: string
        detail:
          type: string
        meta:
          $ref: '#/components/schemas/UnknownErrorMeta'
    UnknownErrorMeta:
      type: object
      properties:
        exceptionClass:
          type: string
        stackTrace:
          type: string
    ErrorInvalidRequestBodyMeta:
      type: object
      properties:
        code:
          type: string
        detail:
          type: string
        meta:
          $ref: '#/components/schemas/InvalidRequestBodyMeta'
    ErrorResponseInvalidRequestBodyMeta:
      type: object
      properties:
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ErrorInvalidRequestBodyMeta'
    InvalidRequestBodyMeta:
      type: object
      properties:
        field:
          type: string
        invalidValue:
          type: object
    ErrorInvalidContentTypeMeta:
      type: object
      properties:
        code:
          type: string
        detail:
          type: string
        meta:
          $ref: '#/components/schemas/InvalidContentTypeMeta'
    ErrorResponseInvalidContentTypeMeta:
      type: object
      properties:
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ErrorInvalidContentTypeMeta'
    InvalidContentTypeMeta:
      type: object
      properties:
        invalidType:
          type: string
        supportedTypes:
          type: array
          items:
            type: string
    ErrorInvalidHttpMethodMeta:
      type: object
      properties:
        code:
          type: string
        detail:
          type: string
        meta:
          $ref: '#/components/schemas/InvalidHttpMethodMeta'
    ErrorResponseInvalidHttpMethodMeta:
      type: object
      properties:
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ErrorInvalidHttpMethodMeta'
    InvalidHttpMethodMeta:
      type: object
      properties:
        invalidMethod:
          type: string
        allowedMethods:
          type: array
          items:
            type: string
    Address:
      type: object
      properties:
        id:
          type: integer
          format: int64
        street:
          type: string
        city:
          type: string
        state:
          type: string
        country:
          type: string
        zipCode:
          type: string
        order:
          $ref: '#/components/schemas/Order'
    Customer:
      type: object
      properties:
        id:
          type: integer
          format: int64
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        orders:
          uniqueItems: true
          type: array
          items:
            $ref: '#/components/schemas/Order'
    Order:
      type: object
      properties:
        id:
          type: integer
          format: int64
        orderTrackingNumber:
          type: string
        totalQuantity:
          type: integer
          format: int32
        totalPrice:
          type: number
        status:
          type: string
        dateCreated:
          type: string
          format: date-time
        lastUpdated:
          type: string
          format: date-time
        orderItems:
          uniqueItems: true
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        customer:
          $ref: '#/components/schemas/Customer'
        shippingAddress:
          $ref: '#/components/schemas/Address'
        billingAddress:
          $ref: '#/components/schemas/Address'
    OrderItem:
      type: object
      properties:
        id:
          type: integer
          format: int64
        imageUrl:
          type: string
        unitPrice:
          type: number
        quantity:
          type: integer
          format: int32
        productId:
          type: integer
          format: int64
        order:
          $ref: '#/components/schemas/Order'
    Purchase:
      type: object
      properties:
        customer:
          $ref: '#/components/schemas/Customer'
        shippingAddress:
          $ref: '#/components/schemas/Address'
        billingAddress:
          $ref: '#/components/schemas/Address'
        order:
          $ref: '#/components/schemas/Order'
        orderItems:
          uniqueItems: true
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
    PurchaseResponse:
      type: object
      properties:
        orderTrackingNumber:
          type: string
    GeoCountry:
      type: object
      properties:
        id:
          type: integer
          format: int64
        code:
          type: string
        name:
          type: string
    GeoState:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        country:
          $ref: '#/components/schemas/GeoCountry'
    PageGeoState:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/GeoState'
        number:
          type: integer
          format: int32
        sort:
          $ref: '#/components/schemas/SortObject'
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: '#/components/schemas/PageableObject'
        empty:
          type: boolean
    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        sort:
          $ref: '#/components/schemas/SortObject'
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        unpaged:
          type: boolean
        paged:
          type: boolean
    SortObject:
      type: object
      properties:
        empty:
          type: boolean
        sorted:
          type: boolean
        unsorted:
          type: boolean
    PageProduct:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/Product'
        number:
          type: integer
          format: int32
        sort:
          $ref: '#/components/schemas/SortObject'
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: '#/components/schemas/PageableObject'
        empty:
          type: boolean
    Product:
      type: object
      properties:
        id:
          type: integer
          format: int64
        category:
          $ref: '#/components/schemas/ProductCategory'
        sku:
          type: string
        name:
          type: string
        description:
          type: string
        unitPrice:
          type: number
        imageUrl:
          type: string
        active:
          type: boolean
        unitsInStock:
          type: integer
          format: int32
        created_at:
          type: string
          format: date-time
    ProductCategory:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string